move code to helper classes to improve readability
