+++ DataTests/Helpers/CookieModelFactory.swift
@@ -1,0 +1,14 @@
+//
+//  CookieModelFactory.swift
+//  DataTests
+//
+//  Created by arthur takahashi on 15/01/21.
+//  Copyright © 2021 Arthur Takahashi. All rights reserved.
+//
+
+import Foundation
+import Domain
+
+func makeCookieModel() -> CookieModel {
+    return CookieModel(fortune: "any_fortune")
+}
