+++ DataTests/Helpers/TextExtensions.swift
@@ -1,0 +1,19 @@
+//
+//  TextExtensions.swift
+//  DataTests
+//
+//  Created by arthur takahashi on 15/01/21.
+//  Copyright © 2021 Arthur Takahashi. All rights reserved.
+//
+
+import Foundation
+import XCTest
+
+extension XCTestCase {
+    func checkMemoryLeak(for instance: AnyObject) {
+        addTeardownBlock { [weak instance] in
+            XCTAssertNil(instance)
+        }
+    }
+}
+
